Abnormal values analysis: "Division by zero" is fixed
The error "Division by zero" is fixed.
For the variant of matches of values P90% and P95% the check for increase (in percentage) of a maximum from P95%.
For the variant of matches of values P5% and P10% the check for increase (in percentage) of P5% from a minimum.
Added flag "check_abnormal_values" to initiate abnormal values analysis (can be "False" by default).