feat: ignore maptypes and explicitly warn user
Because spark MapType doesn't support groupby,
we are unable to easily process and aggregate MapType
data. This push allows us to ignore maptypes and
explicitly warn the user where maptypes are in the data,
as well as inform the user how they can rectify the problem.
This commit also includes tests, as well as a fix for a
division by zero bug

fix: div by zero bug in pandas_profiling.model.spark.table_spark