meteorite example: add dependency workaround
The meteorite example is currently not working.

It gives the following error:

```
cannot import name 'soft_unicode' from 'markupsafe'
```

This seems to be due to markupsafe removing soft_unicode:
https://github.com/pallets/markupsafe/issues/304

We can work around this by pinning to an earlier version of markupsafe.

This gives the following warning:

pandas-profiling 3.2.0 requires markupsafe~=2.1.1, but you have markupsafe 2.0.1 which is incompatible.

but the example seems to work ok.