fix: add handling for spark correlations with no numeric fields
Assembling a vector column in Spark with no numeric columns results in features with
a NULL size, NULL indices, and an empty list of values.
This causes an exception to be raised when computing correlations.

The solution here is to avoid computing the correlation matrix when there are no
interval columns (numeric).

This change addresses issue #1722.